+++ src/app/domain/errors/user-not-found.ts
@@ -1,6 +1,6 @@
 import { DomainError } from './domain-error'
 
-export class UserNotFound extends DomainError {
+export class UserNotFoundError extends DomainError {
   constructor(message?: string) {
     super(message ?? 'User not found.', 404) // Not found
   }