+++ src/app/domain/services/users/create-user.ts
@@ -1,9 +1,9 @@
 import { PgIntegrityConstraintViolation } from '@/db/utils/postgres-errors'
 import { hashPassword } from '@/utils/password'
 import postgres from 'postgres'
-import { EmailOrUsernameAlreadyRegisteredError } from '../errors/email-or-username-already-registered-error'
-import { HashPasswordError } from '../errors/hash-password-error'
 import { insertUser } from '@/db/repositories/user-repository'
+import { HashPasswordError } from '../../errors/hash-password-error'
+import { EmailOrUsernameAlreadyRegisteredError } from '../../errors/email-or-username-already-registered-error'
 
 export type CreateUserInterface = {
   username: string