+++ src/app/domain/services/users/get-user-by-username.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest'
 
 import { makeUser } from '@/test/factories/make-user'
 import { getUserByUsername } from './get-user-by-username'
-import { UserNotFound } from '../errors/user-not-found'
+import { UserNotFoundError } from '../../errors/user-not-found'
 
 describe('get user by username email', () => {
   it('should be able to get user by username', async () => {
@@ -19,6 +19,6 @@ describe('get user by username email', () => {
   it('should not be able to get user with non-existent username', async () => {
     const sut = await getUserByUsername({ username: 'non-existent' })
 
-    expect(sut).toBeInstanceOf(UserNotFound)
+    expect(sut).toBeInstanceOf(UserNotFoundError)
   })
 })