+++ src/app/domain/services/users/get-user-by-username.ts
@@ -1,7 +1,7 @@
 import { db } from '@/db'
 import { schema } from '@/db/schema'
 import { eq, getTableColumns } from 'drizzle-orm'
-import { UserNotFound } from '../errors/user-not-found'
+import { UserNotFoundError } from '../../errors/user-not-found'
 
 type GetUserByUsernameInput = {
   username: string
@@ -16,7 +16,7 @@ export async function getUserByUsername({ username }: GetUserByUsernameInput) {
     .where(eq(schema.users.username, username))
 
   if (!user) {
-    return new UserNotFound()
+    return new UserNotFoundError()
   }
 
   return { user }