+++ src/app/domain/services/users/is-email-available.spec.ts
@@ -4,7 +4,7 @@ import { faker } from '@faker-js/faker'
 
 import { isEmailAvailable } from './is-email-available'
 import { makeUser } from '@/test/factories/make-user'
-import { EmailAlreadyRegisteredError } from '../errors/email-already-registered'
+import { EmailAlreadyRegisteredError } from '../../errors/email-already-registered'
 
 describe('check email', () => {
   it('should be able to check available email', async () => {