+++ src/app/domain/services/users/is-username-available.ts
@@ -1,5 +1,5 @@
 import { getUserByUsername } from '@/db/repositories/user-repository'
-import { UsernameAlreadyRegisteredError } from '../errors/username-already-registered'
+import { UsernameAlreadyRegisteredError } from '../../errors/username-already-registered'
 
 interface IsUsernameAvailableInterface {
   username: string