+++ src/db/repositories/user-repository.ts
@@ -1,10 +1,14 @@
-import type { CreateUserInterface } from '@/app/domain/services/create-user'
+import type { InsertUserModel } from '@/app/domain/entities/user'
 import { db } from '@/db'
 import { schema } from '@/db/schema'
 import { eq } from 'drizzle-orm'
 
 export async function getUserByEmail(email: string) {
   return db.select().from(schema.users).where(eq(schema.users.email, email))
+}
+
+export async function getUserById(id: string) {
+  return db.select().from(schema.users).where(eq(schema.users.id, id))
 }
 
 export async function getUserByUsername(username: string) {
@@ -18,7 +22,7 @@ export async function insertUser({
   email,
   password,
   username,
-}: CreateUserInterface) {
+}: InsertUserModel) {
   return db
     .insert(schema.users)
     .values({
@@ -27,8 +31,4 @@ export async function insertUser({
       password,
     })
     .returning()
-}
-
-export async function getUserById(id: string) {
-  return db.select().from(schema.users).where(eq(schema.users.id, id))
 }